修复了”添加歌手“、”修改歌手基础信息“、”根据专辑获得歌手“和”上传图片“，并添加了”获得一个用户信息”的实现。 （在我的电脑中，./目录是项目的根目录，而非函数所在文件的目录，因此我修改了aliyunossconfig中key.json的文件路径。若实际测试中提示找不到key.json，则可能需要再次修改该路径）

+++ src/main/java/com/example/webmusic/boot/InitSongFiles.java
@@ -231,7 +231,7 @@ public class InitSongFiles {
             // 数据库中不存在符合条件的记录，创建新的记录
             Artist newArtist = Artist.builder()
                     .name(columnValue)
-                    .song_size(1)
+                    .song_size((long)1)
                     .build();
             // 插入新的记录
             artistService.save(newArtist);
